requirements.txt
fastapi==0.115.14
openai==1.93.0
pypdf==5.7.0
python_docx==1.2.0
PyYAML==6.0.2
uvicorn==0.32.0
jinja2==3.1.4
python-multipart==0.0.17